requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=0.8.11
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
rank-bm25>=0.2.2